requirements.txt
langchain_community==0.3.24
langchain_core==0.3.62
langchain_google_genai==2.1.5
python-dotenv==1.1.0
streamlit==1.45.1
